- Initial support for win32
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@654 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/card-etoken.c
@@ -386,7 +386,7 @@ static void parse_sec_attr(struct sc_file *file, const u8 *buf, size_t len)
 	/* acl defaults to 0xFF if unspecified */
 	for (i = 0; i < 9; i++)
 		if (idx[i] != -1)
-			add_acl_entry(file, idx[i], (i < len) ? buf[i] : 0xFF);
+			add_acl_entry(file, idx[i], (u8)((i < len) ? buf[i] : 0xFF));
 }
 
 static int etoken_select_file(struct sc_card *card,
