+++ src/libopensc/card-flex.c
@@ -195,18 +195,18 @@ static int parse_flex_sf_reply(struct sc_context *ctx, const u8 *buf, int buflen
 	}
         p += 2;
 	if (file->type == SC_FILE_TYPE_DF) {
-		add_acl_entry(file, SC_AC_OP_LIST_FILES, p[0] >> 4, is_mf);
-		add_acl_entry(file, SC_AC_OP_DELETE, p[1] >> 4, is_mf);
-		add_acl_entry(file, SC_AC_OP_CREATE, p[1] & 0x0F, is_mf);
+		add_acl_entry(file, SC_AC_OP_LIST_FILES, (u8)(p[0] >> 4), is_mf);
+		add_acl_entry(file, SC_AC_OP_DELETE, (u8)(p[1] >> 4), is_mf);
+		add_acl_entry(file, SC_AC_OP_CREATE, (u8)(p[1] & 0x0F), is_mf);
 	} else { /* EF */
-		add_acl_entry(file, SC_AC_OP_READ, p[0] >> 4, 0);
+		add_acl_entry(file, SC_AC_OP_READ, (u8)(p[0] >> 4), 0);
 		switch (file->ef_structure) {
 		case SC_FILE_EF_TRANSPARENT:
-			add_acl_entry(file, SC_AC_OP_UPDATE, p[0] & 0x0F, 0);
+			add_acl_entry(file, SC_AC_OP_UPDATE, (u8)(p[0] & 0x0F), 0);
 			break;
 		case SC_FILE_EF_LINEAR_FIXED:
 		case SC_FILE_EF_LINEAR_VARIABLE:
-			add_acl_entry(file, SC_AC_OP_UPDATE, p[0] & 0x0F, 0);
+			add_acl_entry(file, SC_AC_OP_UPDATE, (u8)(p[0] & 0x0F), 0);
 			break;
 		case SC_FILE_EF_CYCLIC:
 #if 0
@@ -216,8 +216,8 @@ static int parse_flex_sf_reply(struct sc_context *ctx, const u8 *buf, int buflen
 			break;
 		}
 	}
-	add_acl_entry(file, SC_AC_OP_REHABILITATE, p[2] >> 4, is_mf);
-	add_acl_entry(file, SC_AC_OP_INVALIDATE, p[2] & 0x0F, is_mf);
+	add_acl_entry(file, SC_AC_OP_REHABILITATE, (u8)(p[2] >> 4), is_mf);
+	add_acl_entry(file, SC_AC_OP_INVALIDATE, (u8)(p[2] & 0x0F), is_mf);
 	p += 3;
 	if (*p++)
 		file->status = SC_FILE_STATUS_ACTIVATED;
